Refactor discourse types into centralized types file
