+++ apps/roam/src/utils/getDiscourseRelations.ts
@@ -3,20 +3,12 @@ import type {
   RoamBasicNode,
   TextNode,
 } from "roamjs-components/types/native";
+import type { DiscourseRelation, Triple } from "./types";
+export type { DiscourseRelation, Triple } from "./types";
 import getSettingValueFromTree from "roamjs-components/util/getSettingValueFromTree";
 import toFlexRegex from "roamjs-components/util/toFlexRegex";
 import DEFAULT_RELATION_VALUES from "~/data/defaultDiscourseRelations";
 import discourseConfigRef from "./discourseConfigRef";
-
-export type Triple = readonly [string, string, string];
-export type DiscourseRelation = {
-  triples: Triple[];
-  id: string;
-  label: string;
-  source: string;
-  destination: string;
-  complement: string;
-};
 
 const matchNodeText = (keyword: string) => {
   return (node: RoamBasicNode | TextNode) =>
